% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Draw multiple names in a graph}
\usage{
draw_names(name_vector)
}
\value{
the graph of the two names
}
\description{
Draw multiple names in a graph
}
\examples{
draw_names (c("Daniel", "Pierre"))

}
