% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{This is the function to read multiple excel sheets into a dataframe}
\usage{
multi_excel(file_name)
}
\value{
the list of dataframes coming from the excel tabs
}
\description{
This is the function to read multiple excel sheets into a dataframe
}
\examples{
multi_excel("Werkzaamheden Tom")

}
