% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{read the csv files in a certain folder}
\usage{
my_csv_reader(dir_path = getwd())
}
\arguments{
\item{x}{path to folder}
}
\value{
a list of csv files
}
\description{
read the csv files in a certain folder
}
