% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Function which will save a data frame to a csv file}
\usage{
save_as_csv(df, file_name)
}
\arguments{
\item{df}{dataframe}

\item{file_name}{name of the saved file}
}
\description{
Function which will save a data frame to a csv file
}
\examples{
save_as_csv(iris, "data.iris")
}
