% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Drawing a graph of a name}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{a character}

\item{the_sex}{a character}
}
\value{
The graph of the name in question
}
\description{
Drawing a graph of a name
}
\examples{
draw_a_name("Daniel", "M")

}
